ebook/requirements.txt
streamlit
google-cloud-storage
google-cloud-aiplatform
extra-streamlit-components
moviepy
opencv-python-headless
pytubefix